Module google_translate2::api

Structs

DetectLanguageRequest
The request message for language detection.

DetectionDetectCall
Detects the language of text within a request.

DetectionListCall
Detects the language of text within a request.

DetectionMethods
A builder providing access to all methods supported on detection resources. It is not used directly, but through the Translate hub.

DetectionsListResponse
There is no detailed description.

DetectionsResource
An array of languages which we detect for the given text The most likely language list first.

DetectionsResourceNested
An array of languages which we detect for the given text The most likely language list first.

LanguageListCall
Returns a list of supported languages for translation.

LanguageMethods
A builder providing access to all methods supported on language resources. It is not used directly, but through the Translate hub.

LanguagesListResponse
There is no detailed description.

LanguagesResource
There is no detailed description.

Translate
Central instance to access all Translate related resource activities

TranslateTextRequest
The main translation request message for the Cloud Translation API.

TranslationListCall
Translates input text, returning translated text.

TranslationMethods
A builder providing access to all methods supported on translation resources. It is not used directly, but through the Translate hub.

TranslationTranslateCall
Translates input text, returning translated text.

TranslationsListResponse
The main language translation response message.

TranslationsResource
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.