Struct google_translate2::DetectionMethods

pub struct DetectionMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on detection resources. It is not used directly, but through the Translate hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_translate2 as translate2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use translate2::Translate;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Translate::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.detections();

Methods

impl<'a, C, A> DetectionMethods<'a, C, A>

fn list(&self, q: &Vec<String>) -> DetectionListCall<'a, C, A>

Create a builder to help you perform the following task:

Detect the language of text.

Arguments

q - The text to detect